ai/pyproject.toml
[project]
name = "streaming-ai"
version = "0.1.0"
description = "AI microservice — local LLM, embeddings, semantic search, content analysis"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sentence-transformers>=3.0.0",
    "llama-cpp-python>=0.3.0",
    "huggingface-hub>=0.25.0",
    "anthropic>=0.42.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
